# openapi.yaml
openapi: 3.0.3
info:
  title: Troika Tech Calling Platform API
  description: Production-ready bulk & single calling platform with Exotel integration
  version: 1.0.0
  contact:
    name: Troika Tech
    email: support@troikatech.in
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.troikatech.in
    description: Production
...
paths:
...
  /api/ai/script:
    post:
      summary: Generate AI script (feature flag)
      tags: [AI]
      security:
        - bearerAuth: []
      description: Only available when FEATURE_AI=true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [persona_id]
              properties:
                persona_id:
                  type: integer
                context:
                  type: object
      responses:
        '200':
          description: Script generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  script_blocks:
                    type: array
                    items:
                      type: string
                  compliance:
                    type: array
                    items:
                      type: string
                  engine:
                    type: string
        '503':
          description: AI service unavailable
        '403':
          description: Feature disabled
  /api/ai/summarize:
    post:
      summary: Summarize call (feature flag)
      tags: [AI]
      security:
        - bearerAuth: []
      description: Only available when FEATURE_AI=true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                call_sid:
                  type: string
                recording_url:
                  type: string
      responses:
        '200':
          description: Call summarized
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  key_points:
                    type: array
                    items:
                      type: string
                  sentiment:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...